// src/api/type/persisted-event.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * The schema for persisted events
 */
export interface PersistedEvent {
  event: string
  eventId: string
  /**
   * The timestamp when this event was persisted in the ledger.
   */
  timestamp: string
  /**
   * The key is the name of the entity, value is an array of entity ID values.
   */
  entities: {
    /**
     * This interface was referenced by `undefined`'s JSON-Schema definition
     * via the `patternProperty` "^[^\x00-\x1F\x80-\x9F]{1,80}$".
     */
    [k: string]: string[]
  }
  meta: {
    [k: string]: unknown
  }
  /**
   * Event data can be any type of data, from a string to an object with multiple values. The data is specific to the entity event.
   */
  data:
    | string
    | number
    | boolean
    | {
        [k: string]: unknown
      }
    | unknown[]
}